Switch to bcrypt from argon2.  This will give us a little bit more flexibility if we need it (e.g. interacting directly with PHP or something else).
Added application, group, and apikey objects to the output from users.  This simplifies calls to the API a lot at the expense of more data coming through.  This is only a problem, really, for larger calls where we're returning a large number of users.

Fixed the /users/search call which was returning duplicate objects.